requirements.txt
fastapi
uvicorn[standard]
jinja2
tavily-python
trafilatura
pypdf
requests
sqlalchemy
python-dotenv
aiohttp
langchain
langchain-openai
langchain-core